+++ src/GuiMainMenu.h
@@ -13,6 +13,7 @@ public:
     void onDraw(sf::RenderWindow& wnd);
     void onClick(int button);
     void onLoad();
+    void onResize();
 
 protected: